requirements.txt
allure-pytest==2.11.1
cryptography==40.0.1
filelock==3.12.0
hvac==1.0.2
jsonpath_rw_ext==1.2.2
kafka-python==2.0.1
loguru==0.6.0
mimesis==5.3.0
openpyxl==3.1.2
protobuf==3.20.0
psycopg2-binary==2.9.5
pydantic==1.10.7
pyJWT==2.6.0
pyotp==2.8.0
pytest==7.1.3
pytest-dependency==0.5.1
pytest-xdist==3.1.0
requests==2.28.1
types-requests==2.31.0.1
selenium==4.7.2
sqlalchemy==1.4.22
websocket-client==1.5.1
win32-setctime==1.1.0 ; sys_platform == 'win32'